Fix IncludeFile path regression
Directives processed from IncludeFile are used "as though [they] were
part of the main config.txt file."  Therefore, relative paths given to
the IncludeFile directive are relative to the original config.txt file,
not the subsequent configuration file being processed.  Save the
original config.txt directory as Linter.ConfigDir and use it as the
basis for relative file paths for IncludeFile.

Fixes #55